Integration/requirements.txt
plotly==4.5.1
ibm_cos_sdk==2.0.1
pandas==0.24.2
dash_html_components==1.0.3
nltk==3.4.5
seaborn==0.10.0
numpy==1.18.1
requests==2.22.0
spacy==2.2.4
python_docx==0.8.10
ibm_cloud_sdk_core==1.5.1
wordcloud==1.7.0
watson_developer_cloud==2.10.1
ibm_watson==4.5.0
ibm_cos_sdk_core==2.6.3
dash_core_components==1.10.1
Flask_Cors==3.0.8
Flask==1.1.1
pyresparser==1.0.6
datefinder==0.7.1
matplotlib==3.1.3
dash==1.13.4
tweepy==3.8.0
beautifulsoup4==4.9.1
docx==0.2.4
Pillow==7.2.0
PyGithub==1.51
pywaffle==0.5.1
textblob==0.15.3
git+https://github.com/prowriting/prowritingaid.python.git